requirements.txt
# requirements.txt for FastAPI backend
fastapi
uvicorn
asyncpg
python-dotenv
pydantic
google-genai
mcp
litellm
starlette
google-auth
google-adk